+++ core/src/graphics/graphics_renderer.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 #include <memory>
 #include <mutex>
 #include <vector>
@@ -17,15 +18,15 @@ class IGraphicsRenderer {
   virtual void RenderSprites(
       const std::vector<std::shared_ptr<Sprite>>& sprites) = 0;
 
-  virtual void OnResize(size_t width, size_t height, bool fullscreen) = 0;
+  virtual void OnResize(uint32_t width, uint32_t height, bool fullscreen) = 0;
 
   inline void QueueTextureRelease(IUnknown* texture) {
     std::lock_guard lk(texture_release_queue_mutex_);
     texture_release_queue_.push_back(texture);
   }
 
-  inline size_t get_width() const { return width_; }
-  inline size_t get_height() const { return height_; }
+  inline uint32_t get_width() const { return width_; }
+  inline uint32_t get_height() const { return height_; }
   inline bool is_fullscreen() const { return fullscreen_; }
 
  protected:
@@ -52,12 +53,12 @@ class IGraphicsRenderer {
     return rect;
   }
 
-  inline void set_width(size_t width) { width_ = width; }
-  inline void set_height(size_t height) { height_ = height; }
+  inline void set_width(uint32_t width) { width_ = width; }
+  inline void set_height(uint32_t height) { height_ = height; }
   inline void set_fullscreen(bool fullscreen) { fullscreen_ = fullscreen; }
 
  private:
-  size_t width_, height_;
+  uint32_t width_, height_;
   bool fullscreen_;
 
   std::vector<IUnknown*> texture_release_queue_;
